Add nix-dev script for conditional development environment
This commit introduces a new script, `nix-dev`, which conditionally
executes the `nix develop` command if a `flake.nix` file is present
in the current or parent directories. This allows for a more flexible
development setup by ensuring that the `nix develop` command is only
run when applicable.

Additionally, the tmuxp configuration is updated to use `nix-dev`
instead of `nix develop`, ensuring the tmux session launches the
correct script.

- Added `nix-dev` script to the shared scripts directory
- Updated tmuxp configuration to reference `nix-dev`